parcel_service/requirements.txt
fastapi
uvicorn
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
tortoise-orm==0.19.3